Implement export to .wav
Unlike Bosca, where we let the composition play
out and capture the streamed data, here I decided
to do something more interesting. The reason for
this is both because this method is easier to
implement, and because I wanted to test
functionality which is not used in Bosca.

So we use MML as a medium for rendering out the
sound. The tricky part is getting the timing right.
The rest is pretty simple, as the things that we do
here are easily translated into MML with no loss.

If in the future we add filters, this should also be
trivial to support. And it helped to find some bugs
in GDSiON, so all goals are achieved.