Complete the glow button and integrate it into the layout
For now, the voice type buttons use it, but other
places will rely on it as well. Not all buttons should
have this look, though.

This commit includes a decently sized rewrite of
the shader, since the original attempt didn't work
(well) with arbitrary sizes. We also have transitional
animations now.