Add a warning to the randomizer button
While it's not ideal to scare users like that, it's
better to warn than not to warn here. In a
perfect scenario, all randomized knobs should
not produce any awful, unpleasant sounds, but
I can't guarantee that without excessive testing.

Plus, some combinations may be bad while the
values in individual knobs are still pretty
reasonable. Guardrails for such cases can only
be added manually based on reports. So a
warning (and lower default volume adjusted
earlier) are a good starting point.