Port window manager and settings manager from Bosca Ceoil
This project has similar issue caused by my
hiDPI screen and how it is hard to judge the
scale of the UI for others because of it. The
main issue is that the editor's preview is scaled
with editor's scale, while the runtime is not. So
it's hard to adjust the looks.

I probably want to rescale every part of the UI
in the Glasan's case, something which I wasn't
willing to do with Bosca due to how much
custom UI it has. We have a lot of custom UI
here too, but things are much simpler, I think,
for this option to be considered.

Using downscaling will absolutely destroy a lot
of texts, which would be very unfortunate.