Reduce sizes and margins across the board
This should basically makes the normal-DPI scaling
the baseline, and the hiDPI scaling which I used when
doing the initial layout a modification.

Some implementation notes:
- Clipping is unreliable and whole pixel based in
some way. Because of that rogue pixels can escape
the clipping bounds, which is noticeable with all
scrolling displays (roller knob, option flipper). As a
solution the screen overlay was extended by 1 pixel
in relevant directions to hide those rogue pixels.
This disturbs the look of these widgets, but rogue
pixels are by far worse.