Shrink the roller knob and the voice panel further
The roller knob is the most used control, so it's the
best place to look for extra size. I tried to preserve its
look and feel as much as possible while narrowing
the control down. I think it's good, but maybe some
extra tuning in the future would make it better.